Type Alias massa_db_worker::massa_db::MassaDB

pub type MassaDB = RawMassaDB<Slot, SlotSerializer, SlotDeserializer>;

Wrapped RocksDB database

In our instance, we use Slot as the ChangeID

Aliased Type

struct MassaDB {
    pub db: Arc<DBCommon<SingleThreaded, DBWithThreadModeInner>, Global>,
    pub config: MassaDBConfig,
    pub change_history: BTreeMap<Slot, BTreeMap<Vec<u8, Global>, Option<Vec<u8, Global>>, Global>, Global>,
    pub change_history_versioning: BTreeMap<Slot, BTreeMap<Vec<u8, Global>, Option<Vec<u8, Global>>, Global>, Global>,
    pub change_id_serializer: SlotSerializer,
    pub change_id_deserializer: SlotDeserializer,
    pub current_batch: Arc<Mutex<RawMutex, WriteBatchWithTransaction<false>>, Global>,
}

Fields
db: Arc<DBCommon<SingleThreaded, DBWithThreadModeInner>, Global>

The rocksdb instance

config: MassaDBConfig

configuration for the RawMassaDB

change_history: BTreeMap<Slot, BTreeMap<Vec<u8, Global>, Option<Vec<u8, Global>>, Global>, Global>

In change_history, we keep the latest changes made to the database, useful for streaming them to a client.

change_history_versioning: BTreeMap<Slot, BTreeMap<Vec<u8, Global>, Option<Vec<u8, Global>>, Global>, Global>

same as change_history but for versioning

change_id_serializer: SlotSerializer

A serializer for the ChangeID type

change_id_deserializer: SlotDeserializer

A deserializer for the ChangeID type

current_batch: Arc<Mutex<RawMutex, WriteBatchWithTransaction<false>>, Global>

The current RocksDB batch of the database, in a Mutex to share it

Implementations

impl<ChangeID, ChangeIDSerializer, ChangeIDDeserializer> RawMassaDB<ChangeID, ChangeIDSerializer, ChangeIDDeserializer>where ChangeID: PartialOrd + Ord + PartialEq + Eq + Clone + Debug, ChangeIDSerializer: Serializer<ChangeID>, ChangeIDDeserializer: Deserializer<ChangeID>,

pub fn get_batch_to_stream( &self, last_state_step: &StreamingStep<Vec<u8>>, last_change_id: Option<ChangeID> ) -> Result<StreamBatch<ChangeID>, MassaDBError>

Used for bootstrap servers (get a new batch of data from STATE_CF to stream to the client)

Returns a StreamBatch

pub fn get_versioning_batch_to_stream( &self, last_versioning_step: &StreamingStep<Vec<u8>>, last_change_id: Option<ChangeID> ) -> Result<StreamBatch<ChangeID>, MassaDBError>

Used for bootstrap servers (get a new batch of data from VERSIONING_CF to stream to the client)

Returns a StreamBatch

pub fn write_changes( &mut self, changes: BTreeMap<Key, Option<Value>>, versioning_changes: BTreeMap<Key, Option<Value>>, change_id: Option<ChangeID>, reset_history: bool ) -> Result<(), MassaDBError>

Used for:

Bootstrap clients, to write on disk a new received Stream (reset_history: true)

Normal operations, to write changes associated to a given change_id (reset_history: false)

pub fn get_change_id(&self) -> Result<ChangeID, ModelsError>

Get the current change_id attached to the database.

pub fn set_initial_change_id(&self, change_id: ChangeID)

Set the initial change_id. This function should only be called at startup/reset, as it does not batch this set with other changes.

pub fn set_change_id_to_batch(&self, change_id: ChangeID)

Set the current change_id in the batch

pub fn write_batch_bootstrap_client( &mut self, stream_changes: StreamBatch<ChangeID>, stream_changes_versioning: StreamBatch<ChangeID> ) -> Result<(StreamingStep<Key>, StreamingStep<Key>), MassaDBError>

Write a stream_batch of database entries received from a bootstrap server

pub fn get_xof_db_hash(&self) -> HashXof<HASH_XOF_SIZE_BYTES>

Get the current XOF state hash of the database

fn get_xof_db_hash_opt(&self) -> Option<HashXof<HASH_XOF_SIZE_BYTES>>

Get the current XOF state hash of the database

impl RawMassaDB<Slot, SlotSerializer, SlotDeserializer>

pub fn new(config: MassaDBConfig) -> Self

Returns a new MassaDB instance

pub fn default_db_opts() -> Options

fn new_with_options( config: MassaDBConfig, db_opts: Options ) -> Result<Self, Error>

Returns a new MassaDB instance given a config and RocksDB options